+++ sources/core/objects/Computers.h
@@ -7,6 +7,6 @@ class Computer: public IOnMapObject
 public:
     DECLARE_SAVED(Computer, IOnMapObject);
     DECLARE_GET_TYPE_ITEM(Computer);
-    Computer(size_t id);
+    Computer(quint32 id);
 };
 ADD_TO_TYPELIST(Computer);
